Move REST & V1 api calls to reusable apis folder
